# internal/model/response.go
package model

type ErrorResponse struct {
	Error string `json:"error"`
}

type BalanceResponse struct {
	Current   float64 `db:"current_balance" json:"current"`
	Withdrawn float64 `db:"total_withdrawn" json:"withdrawn"`
}

type AccrualResponse struct {
	Order   string   `json:"order"`
	Status  string   `json:"status"`
	Accrual *float64 `json:"accrual,omitempty"`
}

type WithdrawalResponse struct {
	Order       string  `json:"order"`
	Sum         float64 `json:"sum"`
	ProcessedAt string  `json:"processed_at"`
}